+++ CHANGES.md
@@ -17,6 +17,8 @@ To be released.
  -  Changed `IActionEvaluator.Evaluate(IPreEvaluationBlock)` to
     `IActionEvaluator.Evaluate(IPreEvaluationBlock, HashDigest<SHA256>)`.
     [[#3438]]
+ -  Changed `ActionEvaluator` to accept `IStateStore` instead of
+    `IBlockChainStates`  [[#3439]]
 
 ### Backward-incompatible network protocol changes
 
@@ -34,6 +36,7 @@ To be released.
 
 [#3437]: https://github.com/planetarium/libplanet/pull/3437
 [#3438]: https://github.com/planetarium/libplanet/pull/3438
+[#3439]: https://github.com/planetarium/libplanet/pull/3439
 
 
 Version 3.4.0